Changed step to `long long` type

+++ src/bias/MaxEnt.cpp
@@ -124,8 +124,8 @@ class MaxEnt : public Bias {
   double tstart;
   double tend;
   double avgstep; //current number of samples over which to compute the average. Check if could be replaced bu getStep()
-  long int pace_;
-  long int stride_;
+  long long int pace_;
+  long long int stride_;
   double totalWork;
   double BetaReweightBias;
   double simtemp;
@@ -278,9 +278,9 @@ MaxEnt::MaxEnt(const ActionOptions&ao):
   log.printf("Dumping times for the learning rates are (ps): ");
   for(unsigned i=0; i<tau.size(); i++) log.printf(" %f",tau[i]);
   log.printf("\n");
-  log.printf("Lagrangian multipliers are updated every %ld steps (PACE)\n",pace_);
+  log.printf("Lagrangian multipliers are updated every %lld steps (PACE)\n",pace_);
   log.printf("Lagrangian multipliers output file %s\n",lagmultfname.c_str());
-  log.printf("Lagrangian multipliers are written every %ld steps (PRINT_STRIDE)\n",stride_);
+  log.printf("Lagrangian multipliers are written every %lld steps (PRINT_STRIDE)\n",stride_);
   if(fmt.length()>0)
     log.printf("The format for real number in Lagrangian multipliers file is %s\n",fmt.c_str());
   if(tstart >-1.0 && tend>-1.0)
